Added backend for logger and fixed 2 bugs
"return" expression in scala actually tries to return to the context of the caller. Because the callee actor potentially executes its behaviors in a different context (basically, a different thread/thread pool), an unhandled exception is thrown, which causes each thread of execution to die (remainder: do not use "return" expression in a deferred context of execution).
The second bug was given because the FSM defined by the host received remaining auction messages from clients while it was executing its "host" state, and those messages weren't  handled (yielding a MatchError exception).