WIP: adding user client (with bugs)
The main issue of this actor is that is possible that the actor is blocked reading stdin while the auction ends.
This actor never receives the end message since its blocked in the stdin. We need to add a future call or something
like that to handle that event correctly

Maybe this documentation will help
https://doc.akka.io/docs/akka/current/typed/interaction-patterns.html#request-response-with-ask-from-outside-an-actor