Poor's man future handling for UserClient added and resolved 2 bugs
Reading from stdin is now done in a future fashion, as a way to keep receiving messages and to avoid hanging the dispatcher.
Futures are conditionally reloaded; if an actor spawned a future to read from the console, the user has not yet entered anything and a new value of the item arrives, we do not spawn another future, but check if it was completed. If indeed it was completed, we spawn another future to replace that one. If a future does not resolve and the actor stops executing, it gets cancelled.

A bug was encountered in the parsing of user input. If an invalid null string is entered, the predicate "isValid" yielded True, and the host crashed as "" can't be coerced into a valid integer.

Also, an ultra border case was fixed in the random engine offer generator, in which the budget of the client was just 1 ahead of the item value. In that case, the function call resolves to "between(1, 1)", which yields invalid range and crashes the BotClient.